pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.116.0"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
alembic = "^1.16.2"
bcrypt = "^4.3.0"
passlib = "^1.7.4"
psycopg2 = "^2.9.10"
pydantic = "^2.11.7"
pyjwt = "^2.10.1"
pytest = "^8.4.1"
python-dotenv = "^1.1.1"
python-multipart = "^0.0.20"
sqlalchemy = "^2.0.41"
pytest-asyncio = "^1.0.0"
sqlalchemy-utils = "^0.41.2"
httpx = "^0.28.1"
